Added exception rules, serial data key/alignment
The serial string now begins with a #. This should remove the issue where the
board was receiving the string in the middle. Now the board checks for the #
before continuing to read the string. So far there haven't been the same
issues with data being misread. There were two occurrences where strange
characters were drawn, but those are rare enough that I haven't see the same
behavior.

The python script now deals with (some) exceptions. The ParseError and unclosed
token error seems to be the most popular. For now I've been passing it and
that hasn't caused any issues. In general, most errors/exceptions will just be
passed. I don't particularly care, so long as subsequent tries are succesful.
Perhaps, I'll look into it more later.