Added more serial data verification, moved team names to PROGMEM
There don't seem to be any weird characters showing up anymore. The serial data
received is checked for proper start and end keys and that the data inside
those keys contains acceptable ASCII characters. I figure it's more valuable
or important to skip an update if the received data gets messed up. Game data
changes about every 10-15 seconds (~.1 Hz). Perhaps there could be an ACK
or return byte to signify whether the screen was update. If not then the data
can be immediately resent. Otherwise, the script should continue.

It was determined (correctly, hopefully) that the ton of memory the team name
array took up was the cause of the freezes and related issues. Moving the
names to PROGMEM seems to have taken care of that issue. It looks ugly, but it
works better. Maybe they declarations can be moved to an external or included
file.