Started adding/moving functionality in python script to functions
The main changes here are moving functionality from the while/try loop to
their own functions.

Other changes include a minor edit to how the xml parser is used. The import
statement is different according to the documentation. Also, the various calls
to getchildren were removed as the method is deprecated since 2.7. Lastly, in
doing so hard coded list/element identifiers were removed in favor of actual
tag names.